+++ hexahedron/include/kernel/arch/x86_64/interrupt.h
@@ -89,8 +89,8 @@ typedef struct _x86_64_gdt {
 
 
 // Interrupt/exception handlers
-typedef int (*interrupt_handler_t)(uint32_t, uint32_t, registers_t*, extended_registers_t*);
-typedef int (*exception_handler_t)(uint32_t, registers_t*, extended_registers_t*);
+typedef int (*interrupt_handler_t)(uintptr_t, uintptr_t, registers_t*, extended_registers_t*);
+typedef int (*exception_handler_t)(uintptr_t, registers_t*, extended_registers_t*);
 
 /**** DEFINITIONS ****/
